Improved documentation Improved internal file structure
